worker/requirements.txt
Jinja2==3.0.3
pika==1.2.1
pydantic==1.9.0
django-environ==0.8.1
requests==2.27.1
backoff==1.11.1